requirements.txt
pandas==1.4.3
Pillow==9.2.0
plotly==5.10.0
streamlit==1.12.2
scikit-learn==1.1.1
joblib==1.1.0